Merge pull request #182: Vorbis and Opus fail to play file with large artwork.
Vorbisfile and Opusfile read fully the VorbisComment header. It might contain very large artwork, which causes a problem on my embedded platform as they allocate memory to hold it fully, but that's not an issue here.

Nevertheless, even for regular squeezelite, if the VorbisComment is not fully in the stream buffer, the read callback will return 0 when all data is exhausted and because the streambuffer is locked, file will fail to play with error -133. This PR allow more data to be added to the streambuf while waiting in the read_cb

I also removed some code which I'm pretty sure is useless. This is something I've left when I changed the detection of end of track last year to handle the "last few frames of track would not play". In fact the DECODE_COMPLETE shall simply be sent from the main loop, there is no need to test anything at the beginning, especially testing the fulness of the output buffer is not a good criteria anyway.

-philippe44.

Increase squeezelite revision to 1428.

(cherry picked from commit 74fe7934ec60cc31565f088796f56e911f51679c)

# Conflicts:
#	opus.c
#	vorbis.c